more editing options and exception handling

+++ src/main/java/app/App.java
@@ -29,59 +29,51 @@ public class App {
         this.dao = new BookMarkDAO();
     }
 
-    public App(IO io, BookMarkDAO dao){
-        this.io = io; 
-        this.ui = new TextUI(io); 
-        this.dao = dao; 
+    public App(IO io, BookMarkDAO dao) {
+        this.io = io;
+        this.ui = new TextUI(io);
+        this.dao = dao;
     }
-    
+
     public void run() {
         ui.printWelcomeMessage();
         boolean run = true;
         String command = "";
-        int safety=0;
+        int safety = 0;
         //TODO: hide this to an other class
         while (run) {
             command = ui.getMenuCommand();
-            switch (command) {
-                case ("1"):
-                    Bookmark bookmark = ui.askForBookmark();
-                    if (bookmark != null) {
-                        dao.saveBookmarkToDatabase(bookmark);
-                    }
-                    break;
-                case ("2"):
-                    ui.printBookmarkList(dao.getBookMarksOnDatabase());
-                    break;
-                case ("3"):
-                    String searchfield = ui.askForField();
-                    String search = ui.askForSearch();
-                    ui.printBookmarkList(dao.searchField(searchfield, search));
-                    break;
-                case ("4"):
-                    Long editID = ui.askForEntryToEdit(dao.getBookMarksOnDatabase());
-                    String editfield = ui.askForEditField();
-                    io.println("\nOld values: ");
-                    io.println(dao.getSingleBookmarkInfo(editID));
-                    String newEntry = ui.askForNewField(editfield);
-                    dao.editEntry(editID, editfield, newEntry);
-                    break;
-                case ("5"):
-                    Long bookmark_id = ui.askForEntryToEdit(dao.getBookMarksOnDatabase());
-                    if (bookmark_id != null) {
-                        dao.deleteBookmarkFromDatabase(bookmark_id);
-                    }
-                    break;
-                case ("0"):
-                    ui.printGoodbyeMessage();
+            if (command.equals("1") || command.equals("new")) {
+                Bookmark bookmark = ui.askForBookmark();
+                if (bookmark != null) {
+                    dao.saveBookmarkToDatabase(bookmark);
+                }
+            } else if (command.equals("2") || command.equals("list")) {
+                ui.printBookmarkList(dao.getBookMarksOnDatabase());
+            } else if (command.equals("3") || command.equals("search")) {
+                String searchfield = ui.askForField();
+                String search = ui.askForSearch();
+                ui.printBookmarkList(dao.searchField(searchfield, search));
+            } else if (command.equals("4") || command.equals("edit")) {
+                Long editID = ui.askForEntryToEdit(dao.getBookMarksOnDatabase());
+                String editfield = ui.askForEditField(dao.getSingleBookmarkInfo(editID));
+                io.println("\nOld values: ");
+                io.println(dao.getSingleBookmarkInfo(editID));
+                String newEntry = ui.askForNewField(editfield);
+                dao.editEntry(editID, editfield, newEntry);
+            } else if (command.equals("5") || command.equals("delete")) {
+                Long bookmark_id = ui.askForEntryToEdit(dao.getBookMarksOnDatabase());
+                if (bookmark_id != null) {
+                    dao.deleteBookmarkFromDatabase(bookmark_id);
+                }
+            } else if (command.equals("0") || command.equals("exit")) {
+                ui.printGoodbyeMessage();
+                run = false;
+            } else {
+                ui.printUnrecognizedOption();
+                if (safety++ > 100) {
                     run = false;
-                    break;
-                default:
-                    ui.printUnrecognizedOption();
-                    if(safety++>100){
-                        run=false;
-                    }
-                    break;
+                }
             }
         }
     }
